Configure database per the DatabaseCreatedEvent (#15)
* Configure database per the DatabaseCreatedEvent

The DatabaseCreatedEvent handler took a rather simple approach to
defining database connections and always used a local TCP connection
to the MySQL Router on a standard port. This worked at the time, but
depended on behavior which has since changed. Instead, this patch
changes the _on_database_created event handler to process the endpoints
and configure the slurmdbd service based upon the the endpoints.

The database relation is a generic relation and allows for the slurmdbd
charm to be connected to either a mysql-router charm or to a mysql
service, depending on whether the database service is configured for
high availability or not. The changes in this patch handle both
scenarios, ensuring that the charm properly attempts to configure itself
as directed by the relation data.

In a highly available configuration, the mysql-router subordinate
charm will be configured for connections and the slurmdbd service will
use the unix socket for communication. This involves specifying
environment variables to the slurmdbd process to indicate which socket
to use.

If the slurmdbd charm is related directly to mysql then the slurmdbd
service will connect over a TCP socket. In this case, the slurmdbd.conf
file can be updated to point to the host address and port that the mysql
service is listening on.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

* [chore] Bump slurm-ops-manager version

Bump slurm-ops-manager version to include the changes necessary
for optional StorageHost/StoragePort parameters.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

* Use urlparse to get unix socket URL path

Use the urlparse method to get the file path from the file:// URI
provided as an endpoint.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

* Refactor _update_defaults from SlurmdbdCharm

Move the _update_defaults from teh SlurmdbdCharm to the SlurmdbdManager
class and rename the method from _update_defaults to set_environment_var
which is more accurate to what it does.

Introduce the test_manager.py unit test file to fully cover the
SlurmdbdManager with unit test cases.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

---------

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>